Simplified model version endpoints
